requirements.txt
gradio>=3.0
Pillow
python-barcode
numpy
websocket-client
requests
imageio[ffmpeg]
psutil
pynvml
httpx[socks]
